Update hashmap set method to change capacity
I made the set method update the size of the hashmap when the number of
elements exceeds the load factor of the capacity. The capacity then is
doubled and a new hashmap is created for the elements. Also made a fix
for the set method where I was not updating the value when an existing
key was entered.